Add in logging for when requests do not complete
We should fire the requestIncomplete event for
when there is a delay in returning the data. This
could be because the retrieving class has decided
that it needs extra data. We're logging this for
now since we need to tie it into the retrieve
method so that we can do some progress reporting.