Fix for getTransformToElement failing
It turns out that getTransformToElement will fail whenever you
try and calculate a transform to an element that has its
display set to none. If we check for the attribute being set,
which is admittedly a bit of a hack (instead of properly checking
for its visibility), we can skip over this particular bit of code.