Error events not firing for cached retrieves
If an error occurred during a retrieve, such as when the
data was being parsed on the server side, the error was
not being passed back along to somewhere that could handle
it. The errors were all being silently eaten. This patch
fixes the cached retrieve so that the error is now
passed back.